Move bool and numeric primitive assertion tests to a dedicated test class.
